+++ packages/band/src/bands/BandSurface.tsx
@@ -1,5 +1,4 @@
-import { ReactNode, useMemo } from 'react'
-import { m } from 'framer-motion'
+import { createElement, FC, ReactNode, useCallback, useMemo, useState } from 'react'
 import {
     BandAxisScale,
     getClassName,
@@ -10,6 +9,12 @@ import {
     Y,
     useProcessedData,
     RecordWithId,
+    Rectangle,
+    DataComponent,
+    DataComponentProps,
+    RectangleProps,
+    WithId,
+    useTooltip,
 } from '@chsk/core'
 import { BandSurfaceProps } from './types'
 
@@ -17,6 +22,8 @@ export const BandSurface = ({
     ids,
     expansion = [0, 0],
     interactive = false,
+    tooltip = false,
+    component = Rectangle,
     className,
     setRole = false,
     style,
@@ -24,6 +31,8 @@ export const BandSurface = ({
     const processedData = useProcessedData()
     const { size } = useDimensions()
     const scales = useScales()
+    const { data: tooltipData } = useTooltip()
+    const [active, setActive] = useState<string | undefined>(undefined)
     const horizontal = scales.x.bandwidth() === 0 && scales.y.bandwidth() !== 0
     const indexScale = horizontal ? (scales.y as BandAxisScale) : (scales.x as BandAxisScale)
     const step = indexScale.step()
@@ -37,26 +46,44 @@ export const BandSurface = ({
     const { idSet } = useMemo(() => getIdKeySets(ids, [], processedData), [ids, processedData])
     const compositeClassName = getClassName('bandSurface', className)
 
+    const tooltipId = tooltipData?.data?.[0].id
+    const onMouseEnter = useCallback(
+        (data?: WithId) => {
+            setActive(data?.id)
+        },
+        [setActive]
+    )
+    const onMouseLeave = useCallback(() => {
+        setActive(undefined)
+    }, [setActive])
+    const handlers = interactive ? { onMouseEnter, onMouseLeave } : undefined
+
+    const dataComponent: FC<DataComponentProps<WithId, RectangleProps>> = DataComponent
     const bands: Array<ReactNode> = processedData.data
         .map((seriesData: RecordWithId, j: number) => {
             if (!idSet.has(seriesData.id)) return null
             const indexPos = indexScale(seriesData.id)
             const x = horizontal ? -expansion[0] : indexPos - step / 2
             const y = horizontal ? indexPos - step / 2 : -expansion[0]
-            return (
-                <m.rect
-                    key={keyPrefix + j}
-                    x={x}
-                    y={y}
-                    width={width}
-                    height={height}
-                    role={setRole ? keyPrefix + j : undefined}
-                    className={compositeClassName}
-                    style={style}
-                    initial={interactive ? { opacity: 0 } : undefined}
-                    whileHover={{ opacity: 1 }}
-                />
-            )
+            const tooltipOpacity = tooltip ? tooltipId === seriesData.id : false
+            const interactiveOpacity = interactive ? seriesData.id === active : false
+            const fixedOpacity = !tooltip && !interactive
+            return createElement(dataComponent, {
+                key: keyPrefix + j,
+                component,
+                data: { id: seriesData.id },
+                props: {
+                    setRole,
+                    x,
+                    y,
+                    width,
+                    height,
+                    className: compositeClassName,
+                    style,
+                    opacity: Number(fixedOpacity || interactiveOpacity || tooltipOpacity),
+                },
+                handlers,
+            })
         })
         .filter(Boolean)